Create new persistence variable
Persistence is defined as the # of times the scraper will try to scroll down if it detects that it's still getting the same images.  This is used to detect if you've scrolled to the end.  Higher values make it more robust to slow scrolling but will add an extra delay near the end.  Low values make it more likely to end prematurely but it won't need to wait as long near the end.